Use the error control operator (`@`) instead of `try-catch`
I wanted to set up a static analyzer (namely, larastan [1]) and tried it
on this tutorial repo first. It threw two errors: one was related to the
index, and another one to a "dead catch". That made me think I was doing
not things in the best way possible.

Looking for a better way to handle index errors in PHP, I found this
blog post [2] which suggested the more "graceful approach" of using `@`.
In case that operation fails, it will suppress the error _and assign
`null`_. Comes in handy.

Here's more about that operator from the manual: [3]

[1]: https://github.com/larastan/larastan
[2]: https://chasingcode.dev/blog/php-handle-undefined-index-gracefully/
[3]: https://www.php.net/manual/en/language.operators.errorcontrol.php