Add Docker deployment on GCP with Cloud SQL Proxy
I followed this guide from the Docker docs:
https://docs.docker.com/guides/frameworks/laravel/production-setup/

The repo for those samples, zoomed in the docker directory, is located at:
https://github.com/dockersamples/laravel-docker-examples/tree/main/docker

I removed the local PostgreSQL bits, since I wanted to use Cloud SQL.
It took quite a while. I had to read so many documentation pages and ask
Claude Sonnet 4.5 for a starting point Dockerfile (that I had to tweak
for hours until I understood what I needed to do).

One downside of this is that you need to run it on the Google Compute
Engine for it to connect using the private IP. I could have used the
public IP, but decided against it.

In fact, the entire reason I did this is because I'm tired of just
following tutorials that have bad practices and wanted to Do It Right™.

For some reason, it's only working with database users that don't have
passwords… Strange. I could've figured that out, too, but decided that
having the service account key was, after all, a sufficient protective
measure.